requirements.txt
fastapi
uvicorn
torch
numpy
Pillow
imageio
qdrant-client
ultralytics
clip @ git+https://github.com/openai/CLIP.git
python-multipart
